refactoring to simplify
